requirements.txt
absl-py==0.7.0
imageio==2.5.0
matplotlib==3.0.2
numpy==1.16.0
protobuf==3.6.1
scikit-learn==0.20.2
torch==1.0.0
tqdm==4.31.1
yapf==0.25.0
